+++ tests/test_exact_deduplication.py
@@ -11,11 +11,11 @@ from datasets import Dataset
 from transformers import AutoTokenizer
 
 from axolotl.utils.data import prepare_dataset
-from axolotl.utils.data.constants import ALPACA_MESSAGES_CONFIG_REVISION, SPECIAL_TOKENS
 from axolotl.utils.data.rl import load_prepare_dpo_datasets
 from axolotl.utils.data.utils import deduplicate_and_log_datasets
 from axolotl.utils.dict import DictDefault
 from axolotl.utils.models import load_processor, load_tokenizer
+from tests.constants import ALPACA_MESSAGES_CONFIG_REVISION, SPECIAL_TOKENS
 
 
 def verify_deduplication(actual_dataset, expected_dataset, dataset_name):
